move cpu test helper into sglang.test; drop registered-test exemptions
