+++ test/registered/cpu/test_bmm.py
@@ -4,10 +4,10 @@ import unittest
 # TODO: use interface in cpu.py
 import torch
 import torch.nn as nn
-from utils import precision
 
 from sglang.srt.layers.quantization.fp8_utils import input_to_float8
 from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.cpu_test_utils import precision
 from sglang.test.test_utils import CustomTestCase
 
 register_cpu_ci(est_time=10, suite="base-b-test-cpu")
