+++ test/registered/cpu/test_causal_conv1d.py
@@ -4,9 +4,9 @@ from typing import Optional
 import sgl_kernel  # noqa: F401
 import torch
 import torch.nn.functional as F
-from utils import parametrize, precision
 
 from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.cpu_test_utils import parametrize, precision
 from sglang.test.test_utils import CustomTestCase
 
 register_cpu_ci(est_time=10, suite="base-b-test-cpu")
