+++ test/registered/cpu/test_mamba.py
@@ -4,9 +4,9 @@ import pytest
 import torch
 import torch.nn.functional as F
 from torch.nn.functional import softplus
-from utils import precision
 
 from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.cpu_test_utils import precision
 
 register_cpu_ci(est_time=10, suite="base-b-test-cpu")
