+++ test/registered/cpu/test_mla.py
@@ -2,9 +2,9 @@ import unittest
 
 import torch
 from torch.nn.functional import scaled_dot_product_attention
-from utils import precision
 
 from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.cpu_test_utils import precision
 from sglang.test.test_utils import CustomTestCase
 
 register_cpu_ci(est_time=10, suite="base-b-test-cpu")
