+++ test/registered/cpu/test_moe.py
@@ -16,7 +16,8 @@ prepack = True
 alpha = 1.702
 limit = 7.0
 
-from utils import (
+from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.cpu_test_utils import (
     BLOCK_K,
     BLOCK_N,
     MXFP4QuantizeUtil,
@@ -31,8 +32,6 @@ from utils import (
     torch_w8a8_per_column_fused_moe,
     unpack_and_dequant_awq,
 )
-
-from sglang.test.ci.ci_register import register_cpu_ci
 
 register_cpu_ci(est_time=10, suite="base-b-test-cpu")
