+++ test/registered/cpu/test_norm.py
@@ -3,9 +3,9 @@ from typing import Optional, Tuple, Union
 
 import pytest
 import torch
-from utils import make_non_contiguous, precision
 
 from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.cpu_test_utils import make_non_contiguous, precision
 
 register_cpu_ci(est_time=10, suite="base-b-test-cpu")
 register_cpu_ci(est_time=10, suite="base-b-test-cpu-arm64")
