+++ test/registered/cpu/test_qwen3.py
@@ -2,10 +2,10 @@ import sys
 
 import pytest
 import torch
-from utils import precision
 
 from sglang.srt.utils import is_host_cpu_arm64
 from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.cpu_test_utils import precision
 
 register_cpu_ci(est_time=10, suite="base-b-test-cpu")
 register_cpu_ci(est_time=10, suite="base-b-test-cpu-arm64")
