+++ test/registered/cpu/test_shared_expert.py
@@ -3,7 +3,9 @@ import math
 import unittest
 
 import torch
-from utils import (
+
+from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.cpu_test_utils import (
     BLOCK_K,
     BLOCK_N,
     factor_for_scale,
@@ -15,8 +17,6 @@ from utils import (
     torch_naive_moe,
     torch_w8a8_per_column_moe,
 )
-
-from sglang.test.ci.ci_register import register_cpu_ci
 from sglang.test.test_utils import CustomTestCase
 
 register_cpu_ci(est_time=10, suite="base-b-test-cpu")
