+++ test/registered/cpu/test_spec_kernels.py
@@ -3,10 +3,10 @@ import unittest
 import sgl_kernel  # noqa: F401
 import torch
 import torch.nn.functional as F
-from utils import precision
 
 from sglang.srt.speculative.eagle_utils import TreeMaskMode, organize_draft_results
 from sglang.test.ci.ci_register import register_cpu_ci
+from sglang.test.cpu_test_utils import precision
 from sglang.test.test_utils import CustomTestCase
 
 register_cpu_ci(est_time=20, suite="base-b-test-cpu")
